Refactoring in HttpDownloaderTest and TestWebServerBuilder
